+++ Frieght.Api/Endpoints/LoadEndpoints.cs
@@ -22,7 +22,7 @@ public static class LoadEndpoints
         var groups = routes.MapGroup("/loads")
             .WithParameterValidation();
 
-        
+        #region GetLoad
         /// <summary>
         /// Get all loads
         /// </summary>
@@ -30,20 +30,20 @@ public static class LoadEndpoints
         /// <returns></returns>
         groups.MapGet("/", async (ILoadRepository repository, ILogger<LoggerCategory> logger) => 
         {
-            logger.LogInformation("Getting all loads");
             try
             {
                 var loads = await repository.GetLoads();
-                logger.LogInformation("{0} Loads found", loads.Count());
                 return Results.Ok(loads.Select(load => load.asDto()));
             }
             catch (Exception ex)
             {
-                logger.LogError(ex, "An error occurred while getting all loads");
-                return Results.Problem("An error occurred while getting all loads", statusCode: 500);
+                logger.LogError(ex, "Failed to retrieve all loads");
+                return Results.Problem("An error occurred while retrieving loads", statusCode: 500);
             }
         });
+        #endregion
 
+        #region GetLoadById
         /// <summary>
         ///  Get Load by Id
         /// </summary>
@@ -52,21 +52,21 @@ public static class LoadEndpoints
         /// <returns></returns>
         groups.MapGet("/{id}", async (ILoadRepository repository, int id, ILogger<LoggerCategory> logger) =>
         {
-            logger.LogInformation("Getting Load by Id {0}", id);
             try
             {
-                Load? load = await repository.GetLoad(id);
+                var load = await repository.GetLoad(id);
                 logger.LogInformation("Load found: {0}", load);
-                return load is not null ? Results.Ok(load.asDto()) : Results.NotFound();
+                return load != null ? Results.Ok(load.asDto()) : Results.NotFound();
             }
             catch (Exception ex)
             {
-                logger.LogError(ex, "An error occurred while getting Load by Id {0}", id);
-                return Results.Problem("An error occurred while getting Load by Id", statusCode: 500);
+                logger.LogError(ex, "Failed to retrieve load by Id {Id}", id);
+                return Results.Problem("An error occurred while retrieving the load", statusCode: 500);
             }
-
         }).WithName(GetLoadEndpointName);
+        #endregion
 
+        #region PostLoad
         /// <summary>
         /// Post a load
         /// </summary>
@@ -75,59 +75,81 @@ public static class LoadEndpoints
         /// <param name="carrierRepository"></param>
         /// <param name="messageSender"></param>
         /// <returns></returns>
-        groups.MapPost("/", async (ILoadRepository repository, CreateLoadDto loadDto,
-            ICarrierRepository carrierRepository, IMessageSender messageSender, ILogger<LoggerCategory> logger) =>
+        groups.MapPost("/", async (ILoadRepository repository, CreateLoadDto loadDto, ICarrierRepository carrierRepository, IMessageSender messageSender, ILogger<LoggerCategory> logger) =>
         {
             logger.LogInformation("Creating Load");
             try
             {
-                await repository.CreateLoad(loadDto);
-                logger.LogInformation("Load Created");
-            }
-            catch (Exception ex)
-            {
-                logger.LogError(ex, "An error occurred while creating Load");
-                return Results.Problem("An error occurred while creating Load", statusCode: 500);
-            }
+                var load = new Load
+                {
+                    ShipperUserId = loadDto.ShipperUserId,
+                    Origin = loadDto.Origin,
+                    Destination = loadDto.Destination,
+                    PickupDate = loadDto.PickupDate,
+                    DeliveryDate = loadDto.DeliveryDate,
+                    Commodity = loadDto.Commodity,
+                    Weight = loadDto.Weight,
+                    OfferAmount = loadDto.OfferAmount,
+                    LoadDetails = loadDto.LoadDetails,
+                    LoadStatus = loadDto.LoadStatus,
+                    CreatedAt = DateTime.UtcNow,
+                    Shipper = new User
+                    {
+                        UserId = loadDto.ShipperUserId,
+                        Email = loadDto.CreatedBy.Email,
+                        CompanyName = loadDto.CreatedBy.CompanyName,
+                        DOTNumber = loadDto.CreatedBy.DOTNumber,
+                        FirstName = loadDto.CreatedBy.FirstName,
+                        LastName = loadDto.CreatedBy.LastName,
+                    }
+                };
+
+                var shipper = new User
+                {
+                    UserId = loadDto.ShipperUserId,
+                    Email = loadDto.CreatedBy.Email,
+                    CompanyName = loadDto.CreatedBy.CompanyName,
+                    DOTNumber = loadDto.CreatedBy.DOTNumber,
+                    FirstName = loadDto.CreatedBy.FirstName,
+                    LastName = loadDto.CreatedBy.LastName,
+                    UserType = "shipper"
+                };
+
+                await repository.CreateLoad(load, shipper);
+                logger.LogInformation("Load Created Successfully");
+
+                // TODO: Uncomment the following code to notify carriers and implemente the NotifyCarriers method
+                // This should be done by getting User info from the User table and use them to Notify the carrier.
+
+                // logger.LogInformation("Notifying Carriers");
+           
+                // await NotifyCarriers(carrierRepository, messageSender, load, logger);
+                // logger.LogInformation("Carriers Notified");
             
-            //Notify Carrier of new load posting
-            Load load = new()
-            {
-                UserId = loadDto.UserId,
-                ShipperUserId = loadDto.ShipperUserId,
-                Origin = loadDto.Origin,
-                Destination = loadDto.Destination,
-                PickupDate = loadDto.PickupDate,
-                DeliveryDate = loadDto.DeliveryDate,
-                Commodity = loadDto.Commodity,
-                Weight = loadDto.Weight,
-                OfferAmount = loadDto.OfferAmount,
-                LoadDetails = loadDto.LoadDetails,
-                LoadStatus = loadDto.LoadStatus,
-                Created = DateTime.UtcNow
- 
-            };
-
-            logger.LogInformation("Notifying Carriers");
-            try
-            {
-                await NotifyCarriers(carrierRepository, messageSender, load, logger);
-                logger.LogInformation("Carriers Notified");
+                return Results.CreatedAtRoute(GetLoadEndpointName, new { id = load.LoadId }, load.asDto());
             }
             catch (Exception ex)
             {
-                logger.LogError(ex, "An error occurred while notifying carriers");
-                return Results.Problem("An error occurred while notifying carriers", statusCode: 500);
+                logger.LogError(ex, "An error occurred while creating load");
+                return Results.Problem("An error occurred while creating load", statusCode: 500);
             }
-
-            return Results.CreatedAtRoute(GetLoadEndpointName, new { id = load.Id }, load);
 
         });
+        #endregion
 
+        #region NotifyCarriers
+        /// <summary>
+        /// Notify carriers
+        /// </summary>
+        /// <param name="carrierRepository"></param>
+        /// <param name="messageSender"></param>
+        /// <param name="load"></param>
+        /// <param name="logger"></param>
+        /// <returns></returns>
         async Task<string> NotifyCarriers(ICarrierRepository carrierRepository, IMessageSender messageSender, Load load, ILogger<LoggerCategory> logger)
         {
             logger.LogInformation("Getting Carriers");
-            var carriers = await carrierRepository.GetCarriers();
+            var carriers = await carrierRepository.GetCarrierByUserType("carrier");
             if(carriers is not null)
             {
                 logger.LogInformation("Carriers found: {0}", carriers.Count());
@@ -135,7 +157,7 @@ public static class LoadEndpoints
                 {
                    try 
                    {
-                        await messageSender.SendEmailAsync(carrier.CompanyEmail, $"From {load.Origin} to {load.Destination}", load.LoadDetails);
+                        await messageSender.SendEmailAsync(carrier.Email, $"From {load.Origin} to {load.Destination}", load.LoadDetails);
                         logger.LogInformation("Carrier {0} notified", carrier.CompanyName);
                     }
                     catch (Exception ex)
@@ -146,7 +168,9 @@ public static class LoadEndpoints
             }
             return  "All carriers have been notified";
         }
+        #endregion
 
+        #region UpdateLoad
         /// <summary>
         /// Update load
         /// </summary>
@@ -156,13 +180,18 @@ public static class LoadEndpoints
         /// <returns></returns>
         groups.MapPut("/{id}", async (ILoadRepository repository, int id, UpdateLoadDto updatedLoadDto, ILogger<LoggerCategory> logger) =>
         {
-            logger.LogInformation("Updating Load");
+            logger.LogInformation("Updating Load with ID: {Id}", id);
             try
             {
-                Load? existingLoad = await repository.GetLoad(id);
-                if (existingLoad is null) return Results.NotFound();
-        
-                existingLoad.UserId = updatedLoadDto.UserId;
+                var existingLoad = await repository.GetLoad(id);
+                if (existingLoad == null)
+                {
+                    logger.LogWarning("Load with ID: {Id} not found", id);
+                    return Results.NotFound("Load not found");
+                }
+
+                // Map updatedLoadDto to existingLoad
+                existingLoad.ShipperUserId = updatedLoadDto.ShipperUserId;
                 existingLoad.Origin = updatedLoadDto.Origin;
                 existingLoad.Destination = updatedLoadDto.Destination;
                 existingLoad.PickupDate = updatedLoadDto.PickupDate;
@@ -172,20 +201,22 @@ public static class LoadEndpoints
                 existingLoad.OfferAmount = updatedLoadDto.OfferAmount;
                 existingLoad.LoadDetails = updatedLoadDto.LoadDetails;
                 existingLoad.LoadStatus = updatedLoadDto.LoadStatus;
-                existingLoad.Modified = DateTime.UtcNow;
+                existingLoad.ModifiedAt = DateTime.UtcNow;
                 existingLoad.ModifiedBy = updatedLoadDto.ModifiedBy;
 
                 await repository.UpdateLoad(existingLoad);
-                logger.LogInformation("Load Updated");
+                logger.LogInformation("Load with ID: {Id} updated successfully", id);
                 return Results.NoContent();
             }
             catch (Exception ex)
             {
-                logger.LogError(ex, "An error occurred while updating Load");
-                return Results.Problem("An error occurred while updating Load", statusCode: 500);
+                logger.LogError(ex, "Failed to update Load with ID: {Id}", id);
+                return Results.Problem("An error occurred while updating the load", statusCode: 500);
             }
         });
+        #endregion
 
+        #region DeleteLoad
         /// <summary>
         /// Delete load by ID
         /// </summary>
@@ -197,21 +228,21 @@ public static class LoadEndpoints
             logger.LogInformation("Deleting Load");
             try
             {
-                Load? load = await repository.GetLoad(id);
-                if (load is not null) await repository.DeleteLoad(id);
-                logger.LogInformation("Load Deleted");
+                var load = await repository.GetLoad(id);
+                if (load == null) return Results.NotFound();
+                await repository.DeleteLoad(id);
+                logger.LogInformation("Load Deleted: {0}", id);
                 return Results.NoContent();
             }
             catch (Exception ex)
             {
-                logger.LogError(ex, "An error occurred while deleting Load");
-                return Results.Problem("An error occurred while deleting Load", statusCode: 500);
+                logger.LogError(ex, "Failed to delete load by Id {Id}", id);
+                return Results.Problem("An error occurred while deleting the load", statusCode: 500);
             }
 
         });
+        #endregion
 
         return groups;
     }
 }
-
-