+++ Frieght.Api/Infrastructure/FrieghtDbContext.cs
@@ -12,10 +12,10 @@ namespace Frieght.Api.Infrastructure
         {
 
         }
-        public DbSet<Load> Loads => Set<Load>();
-        public DbSet<Carrier> Carriers => Set<Carrier>();
-        public DbSet<Bid> Bids => Set<Bid>();
-        public DbSet<Shipper> Shipper => Set<Shipper>();
+        public DbSet<Load> Loads { get; set; }
+        public DbSet<User> Users { get; set; } // Unified set for users
+        public DbSet<Bid> Bids { get; set; } // Assuming Bidding instead of Bid
+
         protected override void OnModelCreating(ModelBuilder modelBuilder)
         {
             modelBuilder.ApplyConfigurationsFromAssembly(Assembly.GetExecutingAssembly());