+++ Frieght.Api/Repositories/BidRepository.cs
@@ -16,93 +16,168 @@ public class BidRepository : IBidRepository
         _logger = logger;
     }
     
-    public async Task CreateBid(Bid bid)
+    #region CreateBid
+    /// <summary>
+    /// Create a bid
+    /// </summary>
+    /// <param name="bid"></param>
+    /// <param name="carrier"></param>
+    /// <returns>None</returns>
+    public async Task CreateBid(Bid bid, User carrier)
     {
         try
         {
-            _logger.LogInformation("Creating Bid: {0}", bid);
-            context.Add(bid);
+            using var transaction = context.Database.BeginTransaction();
+
+            _logger.LogInformation("Attempting to create bid for LoadId: {LoadId}", bid.LoadId);
+
+            // Check if the Shipper already exists
+            var existingCarrier = await context.Users
+                            .FirstOrDefaultAsync(u => u.UserId == bid.CarrierId);
+                            
+            if (existingCarrier == null)
+            {
+                _logger.LogInformation("Carrier not found, creating a new one.");
+                if (carrier == null)
+                {
+                    _logger.LogError("Carrier information is missing in the request.");
+                    throw new ArgumentNullException("Carrier", "Carrier information is required to create a new bid.");
+                }
+                context.Users.Add(carrier);
+                _logger.LogInformation("New Carrier created with UserId: {UserId}", carrier.UserId);
+            }
+            context.Bids.Add(bid);
             await context.SaveChangesAsync();
+            await transaction.CommitAsync();
+            _logger.LogInformation("Bid created successfully with BidId: {BidId}", bid.Id);
         }
         catch (Exception ex)
         {
             _logger.LogError(ex, "Error occurred while creating bid");
-            throw ex;
+            throw;
         }
     }
+    #endregion
 
+    #region DeleteBid
+    /// <summary>
+    /// Delete a bid
+    /// </summary>
+    /// <param name="id"></param>
+    /// <returns>None</returns>
     public async Task DeleteBid(int id)
     {
         try
         {
-            _logger.LogInformation("Deleting Bid: {0}", id);
+            _logger.LogInformation("Deleting Bid: {Id}", id);
             var bid = await context.Bids.FindAsync(id);
-            if (bid is not null) context.Remove(bid);
-            await context.SaveChangesAsync();
+            if (bid != null) 
+            {
+                context.Bids.Remove(bid);
+                await context.SaveChangesAsync();
+            }
         }
         catch (Exception ex)
         {
             _logger.LogError(ex, "Error occurred while deleting bid");
-            throw ex;
+            throw;
         }
     }
+    #endregion
 
+    #region GetBid
+    /// <summary>
+    /// Get Bid by Id
+    /// </summary>
+    /// <param name="id"></param>
+    /// <returns>Bid Object</returns>
     public async Task<Bid?> GetBid(int id)
     {
-        try
+       try
         {
-            _logger.LogInformation("Retrieving Bid: {0}", id);
-            return await context.Bids.FindAsync(id);
+            _logger.LogInformation("Retrieving Bid: {Id}", id);
+            return await context.Bids
+                                .Include(b => b.Load)
+                                .Include(b => b.Carrier)
+                                .FirstOrDefaultAsync(b => b.Id == id);                
         }
         catch (Exception ex)
         {
             _logger.LogError(ex, "Error occurred while retrieving bid");
-            throw ex;
+            throw;
         }
     }
+    #endregion
 
-    public async Task<Bid?> GetBidByLoadId(int id)
+    #region GetBidByLoadIdAndCarrierId
+    /// <summary>
+    /// Get Bid by LoadId and CarrierId
+    /// </summary>
+    /// <param name="loadId"></param>
+    /// <param name="carrierId"></param>
+    /// <returns>Bid Object</returns>
+    public async Task<Bid?> GetBidByLoadIdAndCarrierId(int loadId, string carrierId)
     {
         try
         {
-            _logger.LogInformation("Retrieving Bid by loadId: {0}", id);
-            return await context.Bids.FirstOrDefaultAsync(b => b.LoadId == id);
+            _logger.LogInformation("Retrieving Bid by LoadId: {LoadId} and carrier id: {CarrierId}", loadId, carrierId);
+            return await context.Bids
+                                .Include(b => b.Load)
+                                .Include(b => b.Carrier)
+                                .Where(b => b.LoadId == loadId && b.CarrierId == carrierId)
+                                .FirstOrDefaultAsync();
         }
         catch (Exception ex)
         {
-            _logger.LogError(ex, "Error occurred while retrieving bid");
-            throw ex;
+            _logger.LogError(ex, "Error occurred while retrieving bid by loadId");
+            throw;
         }
     }
+    #endregion
 
+    #region GetBids
+    /// <summary>
+    /// Get all the bids
+    /// </summary>
+    /// <returns>IEnumerable<Bid></returns>
     public async Task<IEnumerable<Bid>> GetBids()
     {
         try
         {
-            _logger.LogInformation("Retrieving Bids");
-            return await context.Bids.AsNoTracking().ToListAsync();
+            _logger.LogInformation("Retrieving all Bids");
+            return await context.Bids
+                                .Include(b => b.Load)
+                                .Include(b => b.Carrier)
+                                .AsNoTracking()
+                                .ToListAsync();
         }
         catch (Exception ex)
         {
-            _logger.LogError(ex, "Error occurred while retrieving bids");
-            throw ex;
+            _logger.LogError(ex, "Error occurred while retrieving all bids");
+            throw;
         }
     }
+    #endregion
 
+    #region UpdateBid
+    /// <summary>
+    /// Update a bid
+    /// </summary>
+    /// <param name="bid"></param>
+    /// <returns>None</returns>
     public async Task UpdateBid(Bid bid)
     {
         try
         {
-            _logger.LogInformation("Updating Bid: {0}", bid);
-            context.Update(bid);
+            _logger.LogInformation("Updating Bid: {Bid}", bid);
+            context.Bids.Update(bid);
             await context.SaveChangesAsync();
         }
         catch (Exception ex)
         {
             _logger.LogError(ex, "Error occurred while updating bid");
-            throw ex;
+            throw;
         }
     }
+    #endregion
 }
-
-