+++ Frieght.Api/Repositories/ICarrierRepository.cs
@@ -4,9 +4,10 @@ namespace Frieght.Api.Repositories;
 
 public interface ICarrierRepository
 {
-    Task<IEnumerable<Carrier>> GetCarriers();
-    Task<Carrier?> GetCarrier(int id);
-    Task CreateCarrier(Carrier carrier);
-    Task DeleteCarrier(Carrier carrier);
-    Task UpdateCarrier(Carrier carrier);
+    Task<IEnumerable<User>> GetCarriers();
+    Task<IEnumerable<User>> GetCarrierByUserType(string UserType);
+    Task<User?> GetCarrier(string id);
+    Task CreateCarrier(User carrier);
+    Task DeleteCarrier(User carrier);
+    Task UpdateCarrier(User carrier);
 }