+++ Changelog.md
@@ -4,3 +4,4 @@
 * [Task 39] - Finish Auto-deployent setup for frieght.api and make it accessible.
 * [Task 40] -  Add shipper info into load entity
 * [Task 41] -  AddLogger and a code to create shipper when the first load is created to avoid violating foreign key violation
+* [Task 42] - Refactor shipper and carrier into one entity User